+++ CHANGELOG.rst
@@ -1,3 +1,11 @@
+next version
+------------
+
+Changes
+=======
+
+* Change the blqdv type to blq in datainfo
+
 0.99.0 (2023-08-23)
 -------------------